Share alias status transitions
